Small Deployment API refactoring
