+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/DeploymentManagerStub.scala
@@ -3,11 +3,10 @@ package pl.touk.nussknacker.engine.management.periodic
 import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.process.ProcessName
-import pl.touk.nussknacker.engine.management.FlinkProcessStateDefinitionManager
 
 import scala.concurrent.Future
 
-class DeploymentManagerStub extends DeploymentManager {
+class DeploymentManagerStub extends BaseDeploymentManager {
 
   var jobStatus: Option[ProcessState] = None
 
@@ -29,23 +28,11 @@ class DeploymentManagerStub extends DeploymentManager {
 
   override def deploy(processVersion: ProcessVersion, deploymentData: DeploymentData, processDeploymentData: ProcessDeploymentData, savepointPath: Option[String]): Future[Option[ExternalDeploymentId]] = ???
 
-  override def stop(name: ProcessName, savepointDir: Option[String], user: User): Future[SavepointResult] = ???
-
   override def cancel(name: ProcessName, user: User): Future[Unit] = Future.successful(())
 
   override def test[T](name: ProcessName, json: String, testData: TestProcess.TestData, variableEncoder: Any => T): Future[TestProcess.TestResults[T]] = ???
 
   override def findJobStatus(name: ProcessName): Future[Option[ProcessState]] = Future.successful(jobStatus)
 
-  override def savepoint(name: ProcessName, savepointDir: Option[String]): Future[SavepointResult] = ???
-
-  override def processStateDefinitionManager: ProcessStateDefinitionManager = FlinkProcessStateDefinitionManager
-
-  override def close(): Unit = ???
-
-  override def customActions: List[CustomAction] = List.empty
-
-  override def invokeCustomAction(actionRequest: CustomActionRequest, processDeploymentData: ProcessDeploymentData): Future[Either[CustomActionError, CustomActionResult]] =
-    Future.successful(Left(CustomActionNotImplemented(actionRequest)))
 }
